+++ lib/qsu.dart
@@ -5,5 +5,6 @@ export 'src/crypto.dart';
 export 'src/format.dart';
 export 'src/math.dart';
 export 'src/misc.dart';
+export 'src/object.dart';
 export 'src/string.dart';
 export 'src/verify.dart';