+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `numRandom` method
 - Add `len` method
 - Add `isTrueMinimumNumberOfTimes` method
+- Add `objToQueryString` method
 
 ## 1.0.0 (2024-10-19)